+++ test/converters/html-converter.js
@@ -18,25 +18,23 @@ describe('HtmlConverter', function () {
 
   describe('# Functionality', function () {
     it('should be able to convert the given JSON to HTML using the given template and engine', async () => {
-      const html = await HtmlConverter.convert(EX_1_HTML_INFO, EX_1_DATA);
+      const converter = new HtmlConverter();
+      const html = await converter.convert(EX_1_HTML_INFO, EX_1_DATA);
       isHTML(html).should.be.true;
     });
 
     it('should be able to convert the given Markdown to HTML', async () => {
-      const html = await HtmlConverter.convert(EX_2_HTML_INFO, null);
+      const converter = new HtmlConverter();
+      const html = await converter.convert(EX_2_HTML_INFO, null);
       isHTML(html).should.be.true;
     });
 
     it('should be able to convert the given Pug with front matter to HTML', async () => {
-      const html = await HtmlConverter.convert(EX_3_HTML_INFO, EX_3_DATA);
+      const converter = new HtmlConverter();
+      const html = await converter.convert(EX_3_HTML_INFO, EX_3_DATA);
       isHTML(html).should.be.true;
       html.should.deep.equal(EX_3_OUTPUT);
     });
 
-    it('should be able to convert the given markdown with a layout to HTML', async () => {
-      const html = await HtmlConverter.convert(EX_4_HTML_INFO, null);
-      isHTML(html).should.be.true;
-      html.should.deep.equal(EX_4_OUTPUT);
-    });
   })
 });